Snake Game v2
Project completed: 08 - August - 2023

The MVC (non-strict) pattern has been implemented.
Now there is a main menu with different options.
Added a window showing the controls.
There is now a connection to a Database (MySQL).
A window has been added that shows a ranking that is updated with the DB.
The GamePlay_Model class has an option in the draw method to facilitate display by frames.
Added a text file where you can optionally change the draw method to make it more modular. This could help to better visualize each of the components that are drawn (such as painting the snake, the apple, the board, the score, etc).

There are also multiple changes such as the creation of several classes following the non-strict MVC pattern or adding SQL insert statements that prevent SQL injection (since they are "good practices" ).