Challenges in Python
Lessons completed: 05 - September - 2023

After learning Python up to this point, I've successfully created several small programs and projects to solidify my skills in this lenguage. Among these, I've developed a digital clock, a calculator, and even a game of tic-tac-toe. As my most challenging and final project, I've designed a text editor that enables users to write and save text on their computer, or open and modify existing files and save the changes.

In the upcoming phase of my Python journey, I plan to explore using Python for web development, which will mark the culmination of my summer learning with this versatile language.