QuizMasters Game v0.1
Project in development: 12 - September - 2023 [Part-1] 

In this project, I am developing a Java application using the MVC design pattern, with MongoDB as the database.

Using MongoDB has been challenging because I was not familiar with projects structured with Maven or Gradle. The MongoDB driver offered more options than what I was used to with MySQL, and it required effort and research to identify the necessary dependencies and perform the setup to use this non-relational database.

At this point, I can confidently state that MongoDB functions correctly as the database. It handles insertions and searches seamlessly. Error controls have been implemented to allow the creation of new users and access to the game menu only after user verification and retrieval from the database.

These error controls effectively address user duplication issues, ensuring unique usernames and complete password verification that are also secure.